+++ NEWS.md
@@ -5,6 +5,7 @@
 * Changing slider titles from Transparency to Alpha for `tm_g_gh_boxplot` and `tm_g_gh_spaghettiplot`.
 * Replace `plot_with_height` module with new `plot_with_settings` module.
 * Allow to set a minimum number of observations to render a node in `tm_g_gh_lineplot` and allow to change the font size in the table. Also the previously named argument `font_size` is now `plot_font_size` to distinguish it from the table font size argument.
+* Add option to customize symbol type, line type and color in `tm_g_gh_lineplot`.
 * issues with no news:
 
 # teal.goshawk 0.1.5